refactor: update check exception for delete system

+++ src/main/java/com/sharp/vn/its/management/service/SystemService.java
@@ -1,16 +1,24 @@
 package com.sharp.vn.its.management.service;
 
+import com.sharp.vn.its.management.constants.SortType;
 import com.sharp.vn.its.management.dto.system.SystemDTO;
 import com.sharp.vn.its.management.entity.SystemEntity;
 import com.sharp.vn.its.management.entity.UserEntity;
 import com.sharp.vn.its.management.exception.DataValidationException;
 import com.sharp.vn.its.management.exception.ObjectNotFoundException;
+import com.sharp.vn.its.management.filter.SortCriteria;
 import com.sharp.vn.its.management.repositories.SystemRepository;
+import com.sharp.vn.its.management.repositories.TaskRepository;
+import com.sharp.vn.its.management.repositories.UserRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
 
 /**
  * The type System service.
@@ -36,6 +44,9 @@ public class SystemService extends BaseService {
     @Autowired
     private UserRepository userRepository;
 
+    @Autowired
+    private TaskRepository taskRepository;
+
     /**
      * Gets all systems data.
      *
@@ -52,6 +63,7 @@ public class SystemService extends BaseService {
     /**
      * Gets all systems data by search param and pageable.
      *
+     * @param request the request
      * @return the all systems data
      */
     public Page<SystemDTO> loadAllSystemData(SystemDTO request) {
@@ -87,6 +99,9 @@ public class SystemService extends BaseService {
     public void deleteSystem(Long id) {
         if (id == null) {
             throw new DataValidationException("System id not found");
+        }
+        if(taskRepository.existsBySystemId(id)){
+           throw new DataIntegrityViolationException("Cannot delete system with id " + id +" because tasks are associated with it");
         }
         systemRepository.deleteById(id);
         log.info("System with id {} deleted successfully.", id);
